CRUD DE telefonos y cambios de estilos

+++ MiAppSencha/app/view/empleado/EmpleadoController.js
@@ -69,21 +69,36 @@ Ext.define('MiAppSencha.view.empleado.EmpleadoController',{
         Ext.Msg.confirm("Eliminar empleado","¿Seguro que desea eliminar este empleado?",function(respuesta){
             if(respuesta == "yes"){
                 Ext.Ajax.request({
-                    url: Ext.manifest.url_backend + 'empleados/' + empleado.id,
-                    method: 'DELETE',
-
-                    success: function(response, opts){
-                        Ext.Msg.alert("Ok","Empleado eliminado correctamente", function(){
+                    url: Ext.manifest.url_backend + 'telefonos?empleado=' + empleado.id,
+                    
+                    success: function(response,opts){
+                        var telefonos = Ext.decode(response.responseText)
+                        if(telefonos.length > 0){
+                            Ext.Msg.alert('No Permitido','No se puede eliminar empleados con telefonos asociados')
                             Ext.getStore('empleados-listado').reload();
-                        })
+                        }else{
+                            Ext.Ajax.request({
+                                url: Ext.manifest.url_backend + 'empleados/' + empleado.id,
+                                method: 'DELETE',
+                                success: function(response,opts){
+                                    Ext.Msg.alert('Ok','Empleado eliminado correctamente',function(){
+                                        Ext.getStore('empleados-listado').reload();
+                                    })
+                                },
+                                failure: function(response,opts){
+                                    console.log('El servido fallo con el codigo ' + response.status);
+                                }
+                            })
+                        }
                     },
-                    failure: function(response, opts){
-                        console.log('El servidor falló con el código ' + response.status);
+                    failure: function(response,opts){
+                        console.log('El servido fallo con el codigo ' + response.status);
                     }
                 })
             }
         })
     },
+
 
     reloadStore: function(){
         Ext.getStore('empleados-listado').reload();
